updated the HA dispatch to extend the DefaultHaDispatch
